*Add new Scriptable PostProcessor *Add new handling for layout output *Add support for lineweight output *Cleanup and fixes

+++ src/java/org/kabeja/parser/entities/AbstractEntityHandler.java
@@ -122,7 +122,7 @@ public abstract class AbstractEntityHandler implements DXFEntityHandler {
             break;
 
         case LINE_WEIGHT:
-            entity.setLineWeight(value.getDoubleValue());
+            entity.setLineWeight(value.getIntegerValue());
 
             break;
